\name{preprocessHTT}
\alias{preprocessHTT}

\title{Basic Data Preprocessing}

\description{
This package contains following functions to perform basic data preprocessing:
\itemize{
\item uniqueVar: display number of unique values in each column, and show datatype of each column
\item train_test_split: split dataset to training and testing sets
\item one_hot: perform one hot encoding on selected columns
\item to_ordinal: transfer selected columns from numeric datatype to ordinal datatype
\item standardization: standardize selected columns
}
}

\author{
php2560 project \cr
team name: After School Troll Time
}

